+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/configuration/LogCommonConfiguration.kt
@@ -27,11 +27,8 @@
 
 package com.tencent.devops.log.configuration
 
-import com.tencent.devops.log.jmx.LogPrintBean
-import com.tencent.devops.log.service.BuildLogPrintService
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.AutoConfigureOrder
-import org.springframework.cloud.stream.function.StreamBridge
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.Ordered
@@ -53,14 +50,6 @@ class LogCommonConfiguration {
         }
         return StorageProperties()
     }
-
-    @Bean
-    fun buildLogPrintService(
-        bridge: StreamBridge,
-        logPrintBean: LogPrintBean,
-        storageProperties: StorageProperties,
-        logServiceConfig: LogServiceConfig
-    ) = BuildLogPrintService(bridge, logPrintBean, storageProperties, logServiceConfig)
 
     @Bean
     fun defaultKeywords() = listOf(
