+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/configuration/LogMQConfiguration.kt
@@ -44,6 +44,8 @@ import com.tencent.devops.common.web.mq.EXTEND_RABBIT_ADMIN_NAME
 import com.tencent.devops.log.consumer.LogBatchEventListener
 import com.tencent.devops.log.consumer.LogEventListener
 import com.tencent.devops.log.consumer.LogStatusEventListener
+import com.tencent.devops.log.jmx.LogPrintBean
+import com.tencent.devops.log.service.BuildLogPrintService
 import com.tencent.devops.log.service.LogService
 import org.slf4j.LoggerFactory
 import org.springframework.amqp.core.Binding
@@ -61,6 +63,7 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.AutoConfigureOrder
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication
+import org.springframework.cloud.stream.function.StreamBridge
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.Ordered
@@ -82,6 +85,14 @@ class LogMQConfiguration @Autowired constructor() {
 
     @Value("\${log.rabbitmq.storage.maxConcurrency:#{null}}")
     private val storageMaxConcurrency: Int? = null
+
+    @Bean
+    fun buildLogPrintService(
+        bridge: StreamBridge,
+        logPrintBean: LogPrintBean,
+        storageProperties: StorageProperties,
+        logServiceConfig: LogServiceConfig
+    ) = BuildLogPrintService(bridge, logPrintBean, storageProperties, logServiceConfig)
 
     @Bean
     fun rabbitAdmin(
